Time: 34 ms (46.60%) | Memory: 58.4 MB (52.22%) - LeetSync

+++ 2665-minimum-time-to-repair-cars/minimum-time-to-repair-cars.java
@@ -1,8 +1,8 @@
 class Solution {
     public boolean isPoss(int[] ranks, int cars, long time) {
-        int carsCount = 0;
+        long carsCount = 0;
         for (int rank : ranks) {
-            carsCount += Math.sqrt(time / rank);
+            carsCount += (long) Math.sqrt(time / rank);
         }
         return carsCount >= cars;
     }
